Move Score -> Score\Best
