+++ app/Transformers/API/ScoreTransformer.php
@@ -17,14 +17,14 @@
  *    You should have received a copy of the GNU Affero General Public License
  *    along with osu!web.  If not, see <http://www.gnu.org/licenses/>.
  */
-namespace App\Transformers\Score;
+namespace App\Transformers\API;
 
-use App\Models\Score;
+use App\Models\Score\Best;
 use League\Fractal;
 
 class ScoreTransformer extends Fractal\TransformerAbstract
 {
-    public function transform(Score\Model $score)
+    public function transform(Best\Model $score)
     {
         return [
             'beatmap_id' => $score->beatmap_id,
